refactor: move mainpage functions to separate file
